A Bunch of uintptr_t
A couple of these are odd because nearby asserts have (u32) that might suggest they didn't use it originally (but for matching purposes uintptr_t is just a typedef for u32)

Most don't have that issue, though

+++ include/JSystem/J3DGraphBase/J3DTexture.h
@@ -46,8 +46,8 @@ public:
     void setResTIMG(u16 index, const ResTIMG& timg) {
         J3D_ASSERT(81, index < mNum, "Error : range over.");
         mpRes[index] = timg;
-        mpRes[index].imageOffset = ((mpRes[index].imageOffset + (u32)&timg - (u32)(mpRes + index)));
-        mpRes[index].paletteOffset = ((mpRes[index].paletteOffset + (u32)&timg - (u32)(mpRes + index)));
+        mpRes[index].imageOffset = ((mpRes[index].imageOffset + (uintptr_t)&timg - (uintptr_t)(mpRes + index)));
+        mpRes[index].paletteOffset = ((mpRes[index].paletteOffset + (uintptr_t)&timg - (uintptr_t)(mpRes + index)));
     }
 };
