+++ include/JSystem/JKernel/JKRDecomp.h
@@ -7,7 +7,7 @@
 class JKRAMCommand;
 class JKRDecompCommand {
 public:
-    typedef void (*AsyncCallback)(u32);
+    typedef void (*AsyncCallback)(uintptr_t);
 
     JKRDecompCommand();
     ~JKRDecompCommand();
