+++ src/JSystem/J3DGraphBase/J3DShape.cpp
@@ -64,7 +64,7 @@ void J3DLoadCPCmd(u8 cmd, u32 param) {
 /* 802DD308-802DD344       .text J3DLoadArrayBasePtr__F7_GXAttrPv */
 static void J3DLoadArrayBasePtr(GXAttr attr, void* data) {
     u32 idx = (attr == GX_VA_NBT) ? 1 : (attr - GX_VA_POS);
-    J3DLoadCPCmd(0xA0 + idx, ((u32)data & 0x7FFFFFFF));
+    J3DLoadCPCmd(0xA0 + idx, ((uintptr_t)data & 0x7FFFFFFF));
 }
 
 /* 802DD344-802DD3B4       .text loadVtxArray__8J3DShapeCFv */
