+++ src/JSystem/JKernel/JKRDecomp.cpp
@@ -49,7 +49,7 @@ void* JKRDecomp::run() {
         }
 
         if (command->mCallback) {
-            (*command->mCallback)((u32)command);
+            (*command->mCallback)((uintptr_t)command);
             continue;
         }
