+++ src/d/actor/d_a_npc_hi1.cpp
@@ -981,7 +981,7 @@ BOOL daNpc_Hi1_c::bodyCreateHeap() {
 
     mpMorf->getModel()->getModelData()->getJointNodePointer(m_hed_jnt_num & 0xffff)->setCallBack(nodeCB_Head);
     mpMorf->getModel()->getModelData()->getJointNodePointer(m_bbone_jnt_num & 0xffff)->setCallBack(nodeCB_BackBone);
-    mpMorf->getModel()->setUserArea((u32)(this));
+    mpMorf->getModel()->setUserArea((uintptr_t)this);
 
     return TRUE;
 }
